Only attempt to copy file to database if it exists

+++ src/pharmpy/workflows/databases/local_directory.py
@@ -27,7 +27,8 @@ class LocalDirectoryToolDatabase(ToolDatabase):
         super().__init__(toolname)
 
     def store_local_file(self, source_path):
-        shutil.copy2(source_path, self.path)
+        if source_path.is_file():
+            shutil.copy2(source_path, self.path)
 
 
 class LocalDirectoryDatabase(ModelDatabase):
@@ -38,9 +39,11 @@ class LocalDirectoryDatabase(ModelDatabase):
         self.path = path
 
     def store_local_file(self, model, path):
-        shutil.copy2(path, self.path)
+        if path.is_file():
+            shutil.copy2(path, self.path)
 
 
 class LocalModelDirectoryDatabase(ModelDatabase):
     def store_local_file(self, model, path):
-        shutil.copy2(path, model.source.path.parent)
+        if path.is_file():
+            shutil.copy2(path, model.source.path.parent)
